crag-backend/requirements.txt
# Core Dependencies
PyGithub==2.6.1
uvicorn==0.25.0
fastapi==0.109.2
cryptography==44.0.0
attrs==23.1.0
distro==1.9.0
Pillow==10.1.0
filelock==3.13.1
PyJWT==2.10.1
Deprecated==1.2.15
simplejson==3.19.3
pycparser==2.22
wrapt==1.17.1

# GitHub and Git Integration
GitPython==3.1.40

# Code Analysis
bandit==1.7.5
pylint==3.0.2
black==23.11.0

# Database and Storage
SQLAlchemy==2.0.25
pydantic==2.6.3

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
httpx==0.26.0
markdown==3.5
Pygments==2.16.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Additional dependencies
requests==2.31.0
Jinja2==3.1.3
starlette==0.36.3
python-multipart==0.0.9
pyyaml==6.0.2